libs/discord/framework/pyproject.toml
[project]
name = "casting-discord-framework"
version = "0.1.0"
description = "Discord adapter framework for AI at D3"
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.4.0",
    "python-dotenv>=1.0.1",
    "llmgine",
]

[project.scripts]
discord-live-test = "casting.discord.framework.testing.runner:main"

[project.entry-points."casting.plugins"]
discord_framework = "casting.discord.framework:Plugin"

[tool.uv.sources]
llmgine = { workspace = true }

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "casting.discord.framework"
namespace = true